Ensure that GUI1 widgets are drawn on a DRAW_ALL_EVENT
The DRAW_ALL_EVENT is only used in certain circumstances, when a full
redraw of all widgets/components is desired. The previous behaviour
relied on a DRAW-event following the DRAW_ALL_EVENT, which is not
something that can be relied upon. By calling draw() explicitly it
removes the reliance on the DRAW-event.


Former-commit-id: baa2f9c2d88d2aecb8867099d9f316b55ead4648
Former-commit-id: 63b1f566f4536dab139b2b1cacdbe9c916d22277
Former-commit-id: fbbf16757c9aa7b4727eba179e31f8ac36a80b9b
Former-commit-id: f9e2090bf2910a929a9ecdc9dd80c95f816337a9 [formerly 1dd3967fbeed1d3e543b649fea72263b38158dcd]
Former-commit-id: 958efdf5c5e50cce8d165b31979dfdca28ab3a1a