LoW S3: fixed incorrectly placed labels (bugs #22931, #23776)
I'm not sure why, but it seems the labels need to be initialized in prestart
if using [shift_labels] since I removed the explicit loading of that tag's lua
in a preload event. I'm also not sure why the shift values needed to be updated,
but everything seems to work now, which is what matters.


Former-commit-id: 4d317a2fdbb200fa215d7410cf6c4dd8e8a5886f
Former-commit-id: 39dbf81b178fe0b2dd7800276ded46c8fb61dee9
Former-commit-id: 5fe4971efcda18fca33c6d88760df77d1a17c9f7
Former-commit-id: 9535436dbc3ea649d2cda5be2974af36de0c9a24 [formerly 1ead15e26cfa4f42324b287db3c79e259bd552b0]
Former-commit-id: e8ce2f21f4facac09f48cb7e8b1a1277ce932fcb