Resize action buttons dynamically on the main display
This adds the necessary bells and whistles to resize the GUI1 buttons
used in the main game display. It is rather painful as it involves
re-setting most of the attributes that affect the size.

These changes also adds two new draw-methods to the display class and
gets rid of the default parameters. This is to avoid warnings from
clang since the draw() method comes from a superclass to display now.

The GUI1 button's set_image has been updated to treat the image path
in the same way as it is treated in the constructor. This should not
be a problem since the method is not called from anywhere but the
display class, and that call is added in this commit.


Former-commit-id: b1381b90c7c3ce8952c8b57a647c25a1a5d02d05
Former-commit-id: 4ec2d4674a2c1b88cd00c87246cad706b4e46490
Former-commit-id: b7a36e37ddc98f344d5f551ebac9108a740230ae
Former-commit-id: fa74c344b09c5f221509b637412f010f2e733893 [formerly 34d1d3f93ee5b75d7824dbd71f9dc0dace52f57f]
Former-commit-id: 95a03587bc14cfc1025390e4af09592137c511e5