Ensure that the map-editor still works
This makes the map-editor aware of the draw-layering. Since it's a
subclass of the display-class, but relies on the buttons being
re-created in the superclass I've moved that code into the subclass
instead. This way there won't be any infinite loops with a full redraw
triggering another full redraw.

I have also introduced a small workaround in the GUI1 button widget,
where it would sometimes add the postfix to overlay image names that
already contained the postfix. If the image name ends in the postfix,
the postfix is removed from the name before re-added. This ensures
that the files are successfully found.


Former-commit-id: d282eddb9d786753e223192d107c740ba8bcece5
Former-commit-id: 51fd2ffa5b3d4a0a0f01d2f268bacfc790038c68
Former-commit-id: 737416b91099c9c95a7662a4051a18bc4a01a0b4
Former-commit-id: 31a8caed82a0c6fd694365b3bc09c724d4361b69 [formerly e14c7bb6cbfea50f7077ffbf8a9a4a85504c70d5]
Former-commit-id: 9661993977e589ba888385f8a74586c0ac702f9a