Huge refactor of Lua AI engine
This commit potentially breaks any Lua AI customization, except for external Lua candidate actions.
In practice, though, Lua aspects and goals will probably continue to work for the most part.

- The ai table now has a read_only attribute.
  If true, functions that change the game state will be missing from the table.
  The read_only attribute is false in CA execution and in stages.
  It is true everywhere else.
- Every Lua AI component now supports a [args] subtag.
  The contents of this tag are passed as parameters to the component code.
  This data is immutable; components cannot alter its contents.
  (External Lua candidate actions do not receive this data.)
- Accessing the persistent engine data is now supported in all Lua components.

When calling a Lua component, the Lua engine now passes two parameters:
1. The contents of the [args] tag in the specific component.
2. The contents of the [data] tag in the Lua [engine].

The return value of the [engine] code, if any, is stored for later used.
It will be passed as the third parameter to any other Lua component.
This data can be changed, but will not be saved.
The default engine does not return any such data.


Former-commit-id: 09b25dc26002fe21785b5b323e38262b934f93ec
Former-commit-id: ec4940414deca38df1a2f5ad61e60b57705567de
Former-commit-id: 281c212296478f694c3f87423ebba433cf620074
Former-commit-id: c4033d476a8431dc3fcc9bfb5805be23ef963b04 [formerly 4805ac65bad59256258cf2338b29d5a1f5a5614e]
Former-commit-id: cfe79c7be77de5f295c248a5314293b24d78cc49