Make some transient messages transparent again
The transient messages transparency relies on the restorer
pattern. This enables it for certain transient messages, making sure
that they are drawn properly. Enabling it across all messages will
cause breakage.

This is a bit of a workaround, it should be changed to not rely on the
background restore functionality.


Former-commit-id: 6605b650a77cc52aa3765e5300c4542512eecbcd
Former-commit-id: 52c2664d63087d548d6040724a5ced32401f7c64
Former-commit-id: 48502248bbfc402e177b0666743f61a047c8d043
Former-commit-id: 7c7314afa311e0aa96f910a69120fe76d7976025 [formerly cf82b2acc9d32cfd16500b3f6bd076add503e83d]
Former-commit-id: 9c7c0e5ea82321c973a111d7b6ce4698d69822b4