Make sure bigmap isn't overdrawn by in-game UI
The storyscreen would trigger a full redraw-cycle partially through
displaying bigmap. Due to the implementation of the storyscreen, this
would cause it to have the background overdrawn and never updated. The
fix implemented here is to remove layered-drawing functionality from
the storyscreen and trigger a full redraw manually at the end of the
story screen functionality.


Former-commit-id: c9a45e4531271ba067e0b3d41cd49934dc360ee7
Former-commit-id: 1784fd877a719319517d14c49ce0e561e9ace057
Former-commit-id: 618ba42580d86527fcfffd1ad2ded648781105c0
Former-commit-id: 645d37f077e8cf45c2b9b1591cd0c0be4dde8476 [formerly 42cef5926a295dbf9a5b3f2cad7c646bf85c0cd9]
Former-commit-id: 0150785a39e81d5aaa9c93fe7319c77141308747