tlabel: align text to top of available space
In most cases, the widget size is the size of the text and looks no different
with this change. However, in cases where the label widget area is larger than
the text (such as in Unit Create), it resulted in text being rendered in the
middle of the available space, which looked weird.


Former-commit-id: 4e539f8917ee26a3e7f3efc7ece25f5b08caa6e1
Former-commit-id: b2ffac617d93ff3248048a34e6be75e73823f20a
Former-commit-id: 142c7e9315ff46dfbd3095e0c7f78ca7281733d8
Former-commit-id: 5da4d69691ca07c000ba441c430128e9ceeeaa2c [formerly ef68de8b348e12e0a52aa87f7cb5eab0aa9509b8]
Former-commit-id: fc79486e9916e364dc44c74c876e669b7de6fe31