Fix backstab always showing augmented damage in sidebar
Until now, [damage] (and possibly other) specials which relied on filters involving the opponent would display
in the sidebar as if the augmentation were always active (though when attacking, it worked correctly).
This is now fixed - such specials never show the augmentation in the sidebar now.

In addition, some minor code cleanup (initializing the weapon special adjacency filters outside the loop).


Former-commit-id: 2232615f6c3cae7eee1c8f6503ae735c2f85c451
Former-commit-id: 0e287ceba5c2930ef3893384d83a609aa2b5be1e
Former-commit-id: 590aeb3229c3e0096bfabe5634c3a9aa1cc3d48a
Former-commit-id: cbf0618ed4101ed22ca40be5102bced8fa646822 [formerly 2567ace34a93ab1bc3c2369f8d032588170fc069]
Former-commit-id: 18fee2ac5dd1d495857679e7438793917c8bbc5f