improve map::overlay()
we now read the config into objects at the start of that method, this means
that we don't have to do those lookups inside that loop anymore.

This also allows us to change that parameter of that function from config to
std::vector<overlay_rule> later when we want to use that
function from non-wml code.

Former-commit-id: 87c8b1a866fe44b850a67da99cec2c3e68eab369
Former-commit-id: 9bbacfb65aa054b29bdd8a9687874ebb0babe204
Former-commit-id: 32503badee597de45f9f715bae462c54c3be3d88
Former-commit-id: 9973a869485d495420c145a1a56b125f891d854e [formerly 452cacffe82838f642a481c96f7a3228ed1bc48e]
Former-commit-id: b7dcd26d188d438bfbbfc20fc7f46d330d256d4a