Refactor calls to CVideo::flip()
This adds a timer in the SDL event queue that fires every 20ms in order to add a draw event into the queue.
After the draw event has been processed by all listener, a call to CVideo::flip() is made. All other calls to CVideo::flip() have been removed.

This is the first step towards simplifying the rendering path and making sure the screen is in a consistent state when the flip() call is made.
Having the screen in a consistent state will resolve flickering.

This incidentally also fixes resizing in load screens by always blanking them and doing a full re-render.


Former-commit-id: 34b5272d2d2fd771202eef37babe6142e891acdd
Former-commit-id: a197e2a927bc707ce5c72cba2f06658336e0da36
Former-commit-id: 933525382a9a4e25e48cba4663f845fee6d7530a
Former-commit-id: 44f4a1b31b5dea0d5170f7c8c99081e7fc310915 [formerly 3fe67aea120255d084a0b473eb192ebfad90d611]
Former-commit-id: fdfd64b7859611dc35eadf079ad4eadd4f1103fe