Rename .tpp -> .hpp
Several header files used the .tpp extension, which means that they are not correctly detected as C++ (though this can be overridden and has been for one case in the XCode project). The .tpp extension is one of several intended for private template instantiations, but these headers were topleve headers included in many places.


Former-commit-id: 3345df44bb4bcf312d566cfe026a5ede336ec8e5
Former-commit-id: 6d42b2197456fe87ac3ab25c8d74269a7a59ab9e
Former-commit-id: fe7240e4cbb68be1fe29a5687bfd09dda054e858
Former-commit-id: 680014fd2e7a77bf02150de4421cb1a55482e1f8 [formerly 027826a3496224df8af79aa0eb7e746418e044f3]
Former-commit-id: 3d18d79ae9c3336c543cbba147d4803fcc4501cb