Remove AI interface and fallback stage
This removes the C++ common interface intended for creating new AIs.
Anyone wishing to create a new AI written in C++ should now implement it as a custom stage instead.

This commit also removes the C++ implementation of the Idle AI (but leaves the WML definition of it, which is based on the composite AI)
and the fallback stage which was intended to fall back to other C++ AI algorithms.


Former-commit-id: 842a6382a211f44569de0aba6f43b61a0f1aa9e2
Former-commit-id: 6034254ef99213340984675def600d8e8b029a71
Former-commit-id: b2057e792aa2d337c5bf61ddab5baca4eb2bd00e
Former-commit-id: 1cabf9210915a2b5796a12d09480a71cc0fecc71 [formerly 1b78dd37a0f3e842c24dd73f82c9c5a7f714d23c]
Former-commit-id: 27986d645db5972508eaf18c828f57ef7e8fd9be