Implement hotkey release handling.
Hotkeys are now triggered by a key/button release as well as a press.
This information is passed along as HOTKEY_EVENT_TYPE, which may be
HOTKEY_EVENT_PRESS, HOTKEY_EVENT_RELEASE, or HOTKEY_EVENT_REPEAT (for a
held key repeat).

Currently this should preserve the previous functionality. Existing
hotkeys respond to a press or repeat as they did before and ignore a
release.


Former-commit-id: 3ec322b57ec37ee911eb2be1db76260e960db846
Former-commit-id: 5874a6eaa736f446926814f383bfd50f54a50d69
Former-commit-id: 9ca27d25d2a09a96e5298ab9e2d397e0bd8fab89
Former-commit-id: 00c92458ef030877e69306749d70e65fc7e37537 [formerly 9a1f124f74936994fb44dea2b02a8d6fbe40ca9e]
Former-commit-id: bd247e0268e8fa72ed7cf9a3b1444b35707ad08e