disable c++14 build
Travis fails to build with c++14 because the default installed compiler (or rather the default libstdc++ from gcc 4.8) doesn't support it, i tried to fix it by changing the the install script to install gcc5 but somhow that failed too.

Since noone in our active dev team can fix this, we disable this for now. But i'd be useful if this was reenabled and fixed at some point.

Former-commit-id: 26ff71f228cfeb3a80eeac188dfb1981fea4295f
Former-commit-id: 59e1eead535e6474cbdb935fa15099176c728e1e
Former-commit-id: f418d7d36f36831808c03094b20b8a1662f10a6e
Former-commit-id: 73f3485a4bd87421f3e4ee0ceb1bd562a3edc925 [formerly e17beea87d958bbcad9f43e7e18865da77bcd55b]
Former-commit-id: 27b73c8e5fb3889f10c043c6b0cd5c87c5d4fb74