gui: Fix theme UI buttons missing icons or toggle state upon recreation
This is a blanket fix for the issue I previously tried to fix for
specific cases in the following two commits:

 * fada5316eca20fb6cd50acde4075d45d3a7fb5d2 [formerly 7404cc63b79827ff2ca1b6d7428c45fc1c9b6515] [formerly b3fa3919c619a457ac1ed020be1f7733176c3d29 [formerly 81a0f1b7420dacd394da5706177a3e0565d07389]] [formerly d62233ac46fd7c048d9c00dac4ae365203330783] [formerly ce1c3962e50a3e25bc72f88a0fcef511743055d4] [formerly d818eb6316633084da0db6f7349d598baf9b526f] Call display::draw() before setting theme UI button states
 * 7d7bd2423f7e5dede209ee5f3e32d086b6c94260 [formerly a0ca02eac56101248b38cda2066bde971fbc6e50] [formerly 5eff8bcf4c395aa1c05f06d1d0548a81f5c1e1fd [formerly b68d6dd628460668d9a2a842930e865645ec3341]] [formerly 6998abfc75d4d9f46288463be300058cdfa8cf0d] [formerly eef43f20c080bc1fcb8dc4c81105600c6e9e99b3] [formerly 50fd403b074cfc2e5212939bfba847d44801dcd1] Call set_button_state() from playsingle_controller::init_gui()

This final version of the fix is far superior as it doesn't involve
sloppy guesswork regarding the various display class redraw methods and
their call sites. Instead, we just copy the previous state for each
button during reconstruction regardless of the situation at hand. We
know they are the right buttons because they have the same id strings.

It is a little more involved than I would've liked because it has to
deal with a quirk in gui::button's implementation that becomes more
evident with this fix. The following primarily applies to
gui::button::load_images():

gui::button generates its own built-in image size filename suffix for
the button overlays and frames when the button dimensions are already
known, usually after a method call that sets the button's geometry with
a valid size (i.e. not -1234 x -1234). button::set_overlay() happens to
be such a method, but not all buttons use it to set their initial
overlay, and that's where trouble begins...

 * Some theme UI buttons (mostly in the editor, and the Draw Terrains
   and Draw Units toggles on the minimap area in-game) have their
   overlay icon paths set in the WML, including their _25 size suffix.
   These have their overlays (and frames) set initially set by the
   gui::button ctor, which loads the images and figures out the button
   dimensions in the process (via button::load_images()). The ctor
   always does this unconditionally.

 * At that time, button::load_images() will not add an engine-defined
   suffix because the button dimensions are still unknown until it
   finishes. The images are loaded successfully.

 * Subsequent calls to button::set_overlay() *with the same path* will
   have button::load_images() append its own suffix determined from the
   previously determined button dimensions. Hilarity ensues.

The hotkeys engine (via hotkey::command_executor::set_button_state())
and the theme WML have different notions of which button icons have a
suffix defined externally to gui::button. This results in situations
where restoring the previous icons in display::create_buttons() causes
gui::button to look for inexistent images with duplicate size suffixes
("_25_25"). Although modifying the mainline WML to avoid using those
suffixes in the first place would be the most elegant solution, it also
has the potential to cause issues with UMC themes (esp. forks of the
mainine themes), even if those are relatively rare. Moreover, it counts
as a violation of the stable branch ban on API changes.

In other words, even if this part of the fix is very ugly, it's
ultimately necessary if we are to fix the original bug in a stable
branch without breaking the theme WML API contract. This is also why
everything is in a single commit instead of split in two.

Also note that the aforementioned commits aren't fully obsolete; the
most recent one is actually still required so that the hotkeys engine
gets to set the button icon overlays at least once (for those buttons
which don't have this defined by WML) before running WML events. Any
redundancy arising from keeping them wouldn't hurt anyway.

Finally, all this has been rendered irrelevant in master by Aginor's
refactoring and bug fixes, so this code is a technological dead end
anyway.


Former-commit-id: 91644cba42ec64c7debb79ef816abfcd66109257
Former-commit-id: c1308d5d88da789a05c66ec9051277553dc55505
Former-commit-id: 4b3dfbd96a35b5529ae957d8d001df149388320b
Former-commit-id: 04bb208277c3186e29ae3674e0def9cd45f741ea [formerly a9d4e769fbebb85f754929283fadd531d2e23ed3]
Former-commit-id: 5952509c823b865a822f70ab63e9059501420b7f