fix mode=replace in [set_variables]
http://gna.org/bugs/?24671

There was a problem in the code that made sure that the destiation
config is already avaiable when doing a [set_variables]:

The code consists of 2 parts: the first one (as_range_visitor_base)
determines the destination range operated on (a.b or a.b[n]), while the
second one (the .._h classes) implemented the actual operation.

Before this commit the first part also made sure that the destination
range exists before calling the second part, however this is not the
correct way to do it becasue the first part cannot know what part 2
actuall needs: a 'replace' operation at arr[1] needs arr to have at
least 2 elements while an insert operation in aa[1] only needs arr to
have 1 elements. So fixing the 'insert' function
(https://gna.org/bugs/?24412) broke the 'replace' function.

This commit moves the "make sure that the destination range exists" to
the second part, in order to to this we have to split
variable_as_array_h to a seperate visitor because we cannot  easily
implement the "make sure that the destination range exists" in th second
part in a vit_const case.

Maybe this code can be simplified further, in particular i think that we
should maybe remove the 'as_range_visitor_base' completeley since it now
handles just 4 operations (the 4 [set_variables] modes: replace, insert,
append,  merge) from which only 'replace' really operates on a destination
ranges (meaning for those others a.b is just an alias for a.b[0] while in
'replace' a.b means 'replace the whole range' unlike a.b[0] which means
'replace just the first element').


Former-commit-id: e45a6bdf7738d70f61876a524859bafff6089782
Former-commit-id: bc535081d5b926fb1c6eeca7219ed30e873fb601
Former-commit-id: e04f5ae5de3e0663a71c0252bf53e13041d5610c
Former-commit-id: 79e2bc6808b0d4e8966cd93c3fde6d7c1bb9a0a5 [formerly 441f2269804392bb2a36bd6bb063b269be76f877]
Former-commit-id: 1a0c8e29f33576616654ddce0ac5610659a79b77