Demote spurious network error message to info log level
The following message:

    20160516 16:59:30 error network: Destroing playturn_network_adapter with an non empty buffer, this means loss of network data

More often than not, this comes up when performing normal actions that
involve an unexpected but normal quit from a network game (quitting back
to lobby, closing the game), and the 'error' label makes it seem like a
serious issue when it isn't.

Also fixed typos in it.


Former-commit-id: 2077fc0d12e7f22530ec1cc20ce2757f0ea0dae3
Former-commit-id: 6c81ec5e3cff77a0f2b26f2032cbd8099e5102c8
Former-commit-id: a9c1c083df205cdd29d6619f2b63940b54408185
Former-commit-id: add859c9e00a2a6154b67ddd6c9def1fcdd7a770 [formerly abd86534240907cac40c806bdf604d6b064b3d91]
Former-commit-id: 38fc61c773fdefafdee58546a140a8439cbdf1e7