Improve the control reassignment dialog when a player leaves (#654)
This moves the options for giving control to an ally/observer to the top of the list, labels them more descriptively, and prevents the dialog from appearing during linger mode if there is no next scenario. Also, dismissing the dialog by pressing esc no longer produces a network error, but instead simply defaults to setting the side to idle.

Former-commit-id: 1189b02ed1681f227d473b5987144ab43fce0d50
Former-commit-id: cf71e71d2dfc12a6a87108a681e6162f11d362d3
Former-commit-id: 56ef5339e1ce50725fff06c0cf5b3e23b50f8893
Former-commit-id: 2b4115b18ef1d9c26ed5d96a2de77b9d7ee18d93 [formerly ea3d8c92d3095a14b60744f7cd6fa419a0f578cc]
Former-commit-id: df45a1acbac94dcec8f64a90455cc80f0b4dad29