refactor mp and menu_events code
This adds play_controller::is_networked_mp() and
play_controller::send_to_wesnothd() which are now used to comminucate
with wesnothd during a game instead of using network::send_data and
network::nconnections() directly.

The main intention is to make future changes to this code easier
specially:
1) to have more than one connection during a game, for example to check
for updates
2) to replace the clientsided network code with boost asio.

In order to do that i needed a little change to the chat_handler class
which is used both ingame and in the mp lobby. And to do that i
had to cleanup up the menu_events.cpp file and move huge helper
classes into their own files.


Former-commit-id: 7ad4d114c7f992d06e295106614ad109ea30aa63
Former-commit-id: 2ac27b2d82f3807f3786012af793ff95950cf4fe
Former-commit-id: 060e055e2375ed30b845e1e2671bfc3f80a09375
Former-commit-id: 94fd6865c97c6d8d619dfdab540ca52cbf9e66d6 [formerly 8c032e4049058c1fcca8e6140529edad2e3fbe62]
Former-commit-id: 70e5eebd8ba6af2ed97312baf7a6f25ada45b8d8