use boost asio on clientside wesnothd connection
The main advantage is that the new code is much shorter and easier to
understand then the preivous one.

It currently still used the network::error class on some places

This also removed support for 'ping' packages, the plan is to replace it
with SO_KEEPALIVE

Also this temporaily breaks the gui2 lobby since it still uses the old
network code.


Former-commit-id: 7dc2e471a0c8fc5f0a75ae8f7d6da8e7575a5df8
Former-commit-id: dcb6e2914a11a31527bb84c0208edb3b134c4b25
Former-commit-id: 9d368b91fc3370ec905ad181fd34f9faafa13ed3
Former-commit-id: 1eee0ddb9700547f8e729b4421c5e0c06d9bc12e [formerly 06303be49065e0ac97c38d0a964e12c704e11cec]
Former-commit-id: 9c0401021a248c1e35bd4d6636620c4fde1ca46f