allow to write multiple configs in a network package.
It is now possible to write multiple config objects in a single package,
the server reads them then as one big config. Prevoulsy if people wanted
to do that they had to copy both configs into a new config which is
slow, specially if those configs are very big becasue they contain a
whole scenario or an era.


Former-commit-id: a0bf1c32315be91240d3375b9c2b6475b30e9e3d
Former-commit-id: 9f813346b4a1438a04dec86f6d71a84f2cccfd8f
Former-commit-id: 035ad9ba729acf4ee5402057e303d8d283de2272
Former-commit-id: aecc78b406cbba11208c26317d20cff174693dfe [formerly 0fd4f2d52384243c2525e68b3899170a4bc803cc]
Former-commit-id: 73d561c05cba46eeb7ab8c21219bbd0cfeb24c4a