fix client not sending [leave_game]
this restores some codes that send the [leave_game] that was accidently lost when porting the client cost to boost asio.
I'm still unsure about whether it really makes sense to do it this way, specially that we send [leave_game] after (instead of before) calling enter_create_mode in line 501. (Which is was the pre-boost-asio code did aswell).

Former-commit-id: c15408649a421021b03b10038154e83c37ae1a57
Former-commit-id: 66eb18a2f975fbdc50ce5fcb41b07eaa2dfb7a73
Former-commit-id: e949284de46407436534d29c234e2333e4a03cfc
Former-commit-id: fbf5cd1fbb086a54dc65e604e3bb5d94f499a2c8 [formerly a877af66d570501f563e9ba6caf41f08502b69d4]
Former-commit-id: f389d18b4b50c88f78317be6077f7d854f06128f