allow to special apceisl locations in the map_data
it is now possible to specify locations in the editor that can then be
read with wml (without the create scenario editor mode). These special
locations can be placed like starting positions, and can them be used in
the wml code like:
'location_id=<id>' in standard location filters or 'placement=<id>' in
[unit], i also intend to add a lua interface wesnoth.special_locations
later.

It's currently not possible to have multiple special  locations in the
same hex, the main reason is that i'm unsure about how to integrate that
in the editor ui.


Former-commit-id: 313d0b7091f2d5c239960bdb288bcb8a62e79401
Former-commit-id: ba7b889943de34e9c41d1d7da19b8ad852b9bf67
Former-commit-id: b37aada4e4eb7cd9b9fe16da30e0891518715f43
Former-commit-id: e9474fafcd47255049505d74b8362507b45a9cc2 [formerly 4066a5b8c1e590e5e27685e4776c4920e20e74aa]
Former-commit-id: b98d492c7d20b7c513cacac19aaed6ed3760bb62