select special location in editor palette after creating it
This removes the scroll_top() call in palette_manager::adjust_size()
becasue that made it impossible to scroll to the newly created item
after the "New Location Identifer" dialog was shown.

This changes most function in location_palette to use int instead of
size_t becasue "pos - num_visible_items() + 1" might result in a
negative number.

This also fixes a crash that happend when the "Go To" button was used on
an item that was no starting position.


Former-commit-id: c7c3fcdea6511911ab7717a5ef6bef9aedc9f713
Former-commit-id: 4d3c489f921e2607215fd8aea9bfd9896e3a049c
Former-commit-id: 62d277dd2e3027c0f45fd85eb2859bfe77b41f5a
Former-commit-id: d96941fb42945f470301d80ff5b2d368fb1c7585 [formerly 559655560b01de6bfddce3fd3796a0092787549b]
Former-commit-id: 1e3ff24a63aef9fe074af20b7584cd23011b1111