Fix bug #23712: Mage of Light halo not drawing correctly
The halo is recreated when the mage starts to move, and it's initially
placed in the top-left corner of the screen. On the next frame the game
corrects its position. Correcting the position clears the list of hexes the
halo overlaps, which essentially means "the overlapped hexes aren't known,
update them when the halo is redrawn". However, the game also uses the list
of overlapped hexes to determine when the halo should be invalidated, and
thus redrawn: because the list was empty, the halo was never invalidated,
and it stayed at the top-left corner.

Fixed by placing the halo in the right position to begin with. This commit
includes my previous fixes as well, though, because the code was simply
incorrect before (even if it doesn't break anything any more).


Former-commit-id: 305e2d6ca4bace90053ee3bfc08d4a2f6ef3f6a6
Former-commit-id: 8f43b4f9ffee5a8d42afd4ba9c144650a956ab18
Former-commit-id: 3e9ecfbceeea2dd42f08ae0a1376b835830e667c
Former-commit-id: ae6dd81bc8bd3c20f2e27daff6888381b903472a [formerly cf845ebd5438af0e94cf1e46610898deadc76e80]
Former-commit-id: 4e55a0dc93ea6cf05c6055063b519aa8433e695e