Fix bug #24762: Recreate buttons GUI buttons on a full redraw
This re-creates the main display-gui() buttons on a full redraw to take
into account that the theme expects them to be out of sync. As a
consequence I have also been forced to refactor how said buttons are
managed and introduced a new method to join the same event context
as another component.


Former-commit-id: 67ac79f117400b73ce255aa257510bf29a6b8794
Former-commit-id: 3b1430684d355fff044a1e6d28e93b4ee94fc6e7
Former-commit-id: ea953b5bc0c1e88d9fb70b6a6215374987d7a4f7
Former-commit-id: b0c2bf7cd56f10c7c0c492b1945d0ddf6854a275 [formerly 722501429139883a999d2d5deffa385829b6bc21]
Former-commit-id: 5884852afb71c3f1899ef7f6d550ed0198e26b9e