AOI Fix missing speakers, use new [role] features
For each speaking part, make sure it is assigned to a unit. For parts where the conversation suffers if not delivered, recall a unit or recruit one if none are available. For parts where the missing the line does not appear to adversely effect the conversation, omit the lines if no unit can be found on the board at the time.


Former-commit-id: 6c278d2a312a42eef094af525182c31150d7c307
Former-commit-id: 99bc2eb7e18fa0b1d5a1f6438243be7490b4020b
Former-commit-id: 0283d5c85da131a651e8f4bc2bc7b4c28408dceb
Former-commit-id: 72fba492f987786162b365faf1731736f9168e56 [formerly f49c93a64984aa3e61055f19156df206d7b4e0cf]
Former-commit-id: 037fa127f44b0089a582681181a8be0b5604d47d