replace some victory with scenario_end events.
victory is an unsynced event to its potentially unsafe to change the
gamestate in it, this mostly effects LoW (which can be played as a mp
campaign) where i replaced most victory events.

But it also effects sp becasue it can casue bugs to create units in
unsynced events which are used later (here: in later scenarios), so i
replaced also all sp victory events that do that.

Note that the 'scenario_end event'+'proceed_to_next_scenario filter' has
a slightly different meaning than 'victory event': the second gets fired
when there is a local human side who won the scneario while the first
gets fired if there is any human side who on the scenario. In sp this
is the same since there are no remote human sides, but in mp this is the
reason why the first is synced and the second is not.

We could add a seperate event for this case ("before next scenario" or
similar) for easier use, but i didn't find a good name for
that yet.


Former-commit-id: 47ec846feef3ed032d1ea0bd0482c07d62cbe4b0
Former-commit-id: a50e2117a2703eb22809e736b7b05ff86fd2eac2
Former-commit-id: d6f1bed880527362bdd79642dac8a03e5d4706e4
Former-commit-id: d93e71c89ecb6d0654e6a581bc72b4ad8a27eca6 [formerly 8d42b5fa0b71987273b9cefaef48ad397b7c641f]
Former-commit-id: eb7f9e017db28e11a26a9cd775518665bbe5c916