Remove a performance optimization
The attack prediction code used to skip creating the "A slowed" or
"B slowed" plane if A or B was already slowed. As a result, after the
battle the "summary" array of the combatant had a different meaning from
usual: normally summary[0] means the HP distribution when the combatant is
not slowed, and summary[1] means the HP distribution when the combatant is
slowed. However, if the combatant was already slowed before the battle,
summary[0] was the HP distribution when the combatant is slowed and
summary[1] was empty.

This optimization didn't break anything, thanks to itself: because the
"A slowed" or "B slowed" plane wasn't used, the code fetched the
HP distribution from summary[0] instead of summary[1] and therefore
obtained the correct distribution.

However, my Monte Carlo simulation code requires that summary[0] and
summary[1] have their usual meanings. I decided that the optimization is
too hard to support, and removed it.


Former-commit-id: e4231e75701f311073d4c50d5cfa358a28e60a5e
Former-commit-id: 75fd13c9133217bf0688b4a76c72e49c6ca6c543
Former-commit-id: 4940cf298488c6f6b99297ec676070cedf6fef86
Former-commit-id: 8d2019611a8885910f167c9103c0e12c85da2428 [formerly c24cd6e9c3eebc1a9560f170fd0a2f463c2e1784]
Former-commit-id: 244295a9281209fd4f0448521f2e2ee0ccd832e8