Add a destructor to the context class.
This adds a destructor to the context class to ensure that any joined
event_handlers leave the context before it is deallocated.
It also changes the preferences to explicitly leave the global event
context as it is destroyed instead of when the event handler is
destroyed during _exit().
Also replace std::list.splice with push_back and erase(iterator) in
an attempt to avoid compile errors on older versions of gcc.


Former-commit-id: 88616560d118cb802cf4c8b9970c4ca16715436d
Former-commit-id: bfe6d94d623623f2163b40af1d109015e2f4838f
Former-commit-id: 3c8ff1bd9d7a086b3725a43342e91ea14bb87104
Former-commit-id: ac5c94440a0ad3f0851e21c872937d5b60515329 [formerly 13f15e94023cd0c521311ff9d26a4a92ca14aa25]
Former-commit-id: 2efb69e9ff26bb148366e4d0e3c819205936ad3b