make victory event fire on proceed_to_nextlevel
This changes the behaviour of victory event so that they are now also fored when there is at least one human that is victorious, for sp and for coop mp scenarios the conditions when victory is fired is the same as before, in non-coop mp games the victory event will now also be fired whenever there is any human side that wins, since such scenarios usually don't use victory events this is not tha much of an issue. For those cases i added the 'local_victory' event (maybe if people want to print congratulation messages in their mp maps or similar)

The reason for this change is that now, since victory events fire on all client the same ,way we can mark them as synced, this specially fixes issues with invalid unit ids of units created suring victory events in sp.

Former-commit-id: 3b6773de92ac21b70b54bd7e229dd87b7cb4988a
Former-commit-id: b233a6df62f05e6fd965e56622ef9fcb0a2feb2a
Former-commit-id: 9e6a40da120de502c6d86e0f71997303732fe4ce
Former-commit-id: f8f13a7c92b5858539c1f3ec60a12104e9d471c3 [formerly 473c1c8159051b167a4637587838d4f131ec8a9c]
Former-commit-id: 8d2700521c86b4506a8b091eede35d2a7ba50bf6