New features for [role]
Add search_recall_list=only to search the recall list, but not the map
This is intended as a more clear way of doing 'x,y=recall,recall' and omitting search_recall_list. Note that search_recall_list=no with x,y=recall,recall produces a useless [role] which only runs the [else] block(s).

Add reassign attribute
Optional attribute, default 'yes'. If 'no' then do not re-assign the role if a unit already has the role. If 'no' and no unit has the role, the role is still assigned.

Refactor auto_recall as a sub-tag
Instead of a yes|no attribute, if there is an [auto_recall] sub-tag then recall the unit from the recall list. This acts as a [recall] tag without any StandardUnitFilterto which we add id=$unit.id and issue the recall.


Former-commit-id: 59f7fa4c34a1d314f6bc51cdda7e75146164befa
Former-commit-id: 1f1e0b986fd1c6c3c7085d49489d79f966afdd9c
Former-commit-id: 73af68963548b59a53c5e3e34ce94b3a5b7f8c85
Former-commit-id: 5cfdc4497790a4dfa2ab2e1a4f829047b832da75 [formerly 5978f8bfab471d050cf2e83fb717e07a2423f1bd]
Former-commit-id: 9ed9a506ed160f9bbcc6fd9a0f9d3bb1e9ef83a9