Fix undefined behavior on destroying an event context
The destructor of the context class accidentally incremented the iterator
twice per iteration. If the number of event handlers was odd, the
destructor ended up incrementing the end iterator, which is UB.

I rewrote the whole destructor. It's unnecessary to manually remove event
handlers from the list because the list will do it automatically when it's
destroyed.


Former-commit-id: d0f67a35592bdbc1748b937bb4e20d9f94259e0c
Former-commit-id: 0c7350c2897f0396c844ba947e476fd9155a6cf8
Former-commit-id: a53bd36d1035cee0ee0557373fd2caa6a7d55657
Former-commit-id: 6256465d738c62c1f08b398d0432f50539c1b815 [formerly 4636f0a6efef2f21b7afc7134442274981a3b843]
Former-commit-id: 3e0ec5c85f487e16fab3d9a4031916ae0fcbe439