DM S07 Fix bug: Always early completion
The skeleton generators turned off on turn 12, always leading to an early-completion on turn 13. Don't turn them off. This allows normal turns-over completion after 14 turns, as designed.

Note than if the generators are taken out, and all skeletons killed, the player must wait for the next turn for victory. This can be rationalized by Delfador's statement indicating he is not sure his solution will work until the next turn, when no skeleton appears.

The use of an invalid tile to record the state of each generator caused errors generating the preview minimap when loading a turn.  Changing the tiles also caused visual artifacts. Eliminated the custom tiles and use event management to handle the generator state.


Former-commit-id: 8fce9e1baf8740406ed9a6325f5ab147774cff19
Former-commit-id: 776fb86a998e743000eaaae43e12b4bacb559110
Former-commit-id: 9938a59bb30a4b72532d1eb85b917c5f3acea649
Former-commit-id: 3e93889540e40c9ace99281edc5e219338b8808d [formerly 7cfe4e72f2d4429966d6899f25afb1ccf7b81c5e]
Former-commit-id: 6c9e714e049429a4cfb3f119912658688047144b