DM Fix bug: Subsumed sides cannot move
In S13 we subsume the Loyalists and Mages.

In S19 we subsume Lionel and his troops.

These occur in the midst of battle. The units may be damaged, poisoned, etc. We do not want to change that.

However, since their turn follows ours, they cannot move. Reset their moves to maximum so we can use them (or move them out of harm's way) immediately.


Former-commit-id: 7587d995c405cabc65b1956c0aa58420f80a1c6c
Former-commit-id: 202ef5f6fd025cb6cd4ef3c3598f8aba6eec632b
Former-commit-id: c8995246e62996dc294b261e3d67186e86e52d68
Former-commit-id: 1c70298c787538ad5618299db0c4e6d6510c1b70 [formerly c6d24ebb0b4b8f693b4afd72149d91d0a3333d5b]
Former-commit-id: e3c476db2f7bcb044651a4271a39886f52b837f5