Fix crash by reintroduce splice.
This fixes a crash that would be caused by memory incorrectly being freed
when removed from a list. Relying on splice stops that from happening.

This also fixes an exit condition when deleting handlers as depth could
sometimes not be decremented.

Finally, checks are made to made sure that an sdl_handler is either
exclusively in the global context or a different context, ensuring
that they cannot be in both.


Former-commit-id: 8bfb68897d8069b1619d1df82b970b79b0f1101d
Former-commit-id: af9064356eca954845792b1d5c02fe11b9813ca4
Former-commit-id: afab598c51f1dbcd4abb6480d5f334ec4ab70c3c
Former-commit-id: 2068743f184b2da72c9b865278c63557562e597f [formerly b0cf49d47442133883d92c84aa9251a642b7f4cf]
Former-commit-id: 4139a33f9810cc91267fd1e28ac84490f5d46607