Tweaked minimal slider magic numbers again so maximum value can be reached (bug #24531)
There seems to possibly be an internal issue with the slider (and more generally, the
scrollbar implementation form which it inherits) where is min/maximum_positioner_length
equal an even number, sliders with an even number of options will reach max value.

Setting it to what, in this case is 1 below the actual width of the slider image (15)
means both even and odd-numbered sliders will reach max value, however the edge is cut off.

Attempting to set the right_offset to 1 and using the above off value for the positioned
width means only sliders with an odd number of options reach max value.

The magic numbers in this commit seem to give the widget enough space to reach the end of
the slider and max values regardless of the number of options, but the internal widget dimension
calculation should perhaps be investigated in the future.


Former-commit-id: 15db6da2cd0c6210991000b2530a97e08fcd9bac
Former-commit-id: 6748a3c1a8b049352b01ac030df691e26acae916
Former-commit-id: c3da8b38874bfad89873fe88349327f82533278a
Former-commit-id: 293b5865b759b5383f93203d8734e577214bee08 [formerly 82ea00d412817e08d4e7236ae995f13a8b6c2bd1]
Former-commit-id: d68bbe8d361700e4669c57b49948ba8e42efd649