play_controller::check_victory(): use underscore in event name
Semantically this doesn't change anything because spaces in event names are
automatically replaced with underscores.

I just spent a considerable amount of time searching for code that fires
the enemies_defeated event, only to find that it's called
"enemies defeated", with a space, at the call site. Let's use the
underscored name everywhere to make the function easier to find.


Former-commit-id: e35c263cfc99e3ebdf9ab568fce8e6a8faaa7126
Former-commit-id: ee7b3c7aefe13eefc5faa15cdae02617bee58dd4
Former-commit-id: 96566652d746032c9eb06f2058ea92a1b21b7b29
Former-commit-id: 59af947badacb9b2a98e2671f84c2b668a6bbfb8 [formerly a6ce49cdb7b495fe173bb3a4bfb4ea22b016e6b8]
Former-commit-id: 8cfa681029c13d542765451bd8c0833cda950d6b