Fix scenario end events not being run on debug-killing the enemy leader
If a developer used the "kill unit" debug command to end a scenario,
the scenario did something special on victory (e.g. showing messages)
and the scenario didn't explicitly set victory_when_enemies_defeated=no,
the special actions were skipped.

The cause was a reference with a too long lifetime.
Menu_handler::kill_unit() obtained a reference to the mouse position and
used it to send "last_breath" and "die" events. But if the mouse position
changed in between (for example, as a result of the last_breath event
doing something), the reference switched to point to the new mouse
position. As a result, the function sent the "die" event with incorrect
coordinates, filtering dropped the event, and the event handlers weren't
run.

Fixed by copying the mouse position instead.


Former-commit-id: c59494b5c26e6403d65c8998354ebbcf79b67018
Former-commit-id: c404b7c8916e5fd99431c8762b349bc852dd2fa4
Former-commit-id: eba5efd56ab0b3b957bd5f6193d06482559fc932
Former-commit-id: 0c145ed438953f5c99964612cc4ed4f604652207 [formerly 2696af659eeb50576905277b1a48c20b6fd7edd1]
Former-commit-id: 47af4d439cb383fc4ee3699b41cc9f8c4dde0278