Fix bug: libpng warning: iCCP: known incorrect sRGB profile
Some changes in libpng 1.6+ cause it issue warning or even not work correctly with the original HP/MS sRGB profile. You will notice this in stderr:

    libpng warning: iCCP: known incorrect sRGB profile

The old profile uses a D50 whitepoint, where D65 is standard. This profile is not uncommon, being used by Adobe Photoshop, although it was not embedded into images by default.

The simplest solution is simply to remove the embedded profile from your image. This can cause a slight shift in color *IF* have have a properly calibrated system, monitor, and software.


Former-commit-id: c459b22fe3434e25fb25f3e0316caa923934352c
Former-commit-id: 50d812910ac289ddabe041543ff17c6557dc366b
Former-commit-id: 5a52cb374b3322189160ab9be617ba07d2351a4c
Former-commit-id: 3ecdbd7f94c1ee7e179c55993e93481f99ce1a4c [formerly 731b8135faa3a64c6acddd36f5121826a1a5491c]
Former-commit-id: 7642854eef128697543d1918e92d6c352eb19285