HttT S05b Randomize temples
Formerly, to find Moremirmu, simply check (11,13) and, if not there, he is in (10,17). Oh HARD, there was a 1-in-2 chance of finding Xakae in (11,13), instead, and (9,15) was always empty. On NORMAL, Xakae is in (9,15).

Now, the odds of finding Moremirmu in a given temple are 1-in-3, as are the odds of finding Xakae unless playing on EASY where he is not present.


Former-commit-id: 216076425e63309fa99777123df903da8eaae03d
Former-commit-id: 6b974dee47c317116a54ba79e902a1a39bd1ae18
Former-commit-id: 9f30fcfa293110b7ba733018e8fdf9380f29dd5e
Former-commit-id: 5ab5c96e36f58eba0f30ddea8133f3caf678c739 [formerly e7a0f3d759102706da5b2e6ce7fce6b32e9ebd31]
Former-commit-id: 3c19a5b243ac9c5711f25c84dd7cd9c384669041