HttT S07 Fix bug: Units not hiding
Placing a duplicated unit causes it to drop into an adjacent hex and being missed by the object.

This only happens rarely, and almost always at the overlaped column. But it can happen anywhere. We'll just skip it and so what if we're a unit shy. Call it a very tiny, quite rare Easter Egg.


Former-commit-id: d7634372a1d75ec0ecee0dd6b0190ac4d2e04871
Former-commit-id: 503735f41f244aef2d9c8fe1003ce0fc33c7088d
Former-commit-id: 14833597289e012ed6442c3a34b79060cc79e740
Former-commit-id: c40f3900f937e0c14a0cd15e63d5db527b4f0a25 [formerly 0b99142379e0992f6c86474bdea347028f6b5b19]
Former-commit-id: 834fc33d718f2885dbd79b38e8b1cab392698d7b