HttT S13 Add some random snow
The amount of snow is based upon how quickly you made it through S12 (Northern Winter). It is 1/3rd the coverage from S12, so it varries from 0 to 33% total coverage of the map, since we don't want to really bog down the player. At 1/3rd coverage, if S12 was won on the last possible turn the player here should only lose a turn or two getting to the exits.


Former-commit-id: 5a95caaf0c5232a30016ff68b8d21f36b33b37c1
Former-commit-id: 4b489bf14820819e3ae96f214e787ee4084ae57a
Former-commit-id: c1179fa0ca240ef78dc9fe85fa2352e32d0c026f
Former-commit-id: 5f39edfaf95e5791aa84729965951810ea619109 [formerly f09ff5e79cf52f1051a201b899a4890a6957ab73]
Former-commit-id: 3fbc2a19ff3efe1f5e8f0fc9aa3bc00b5eedf587