Damage calculation dialog: use existing battle prediction
The game already predicts the outcome of all possible battles when the
player opens the unit attack dialog, for the purpose of choosing the best
weapon to attack. Simply use that data in the damage calculation dialog
instead of recalculating it.

The change in mouse_handler::fill_weapon_choices() ensures that the damage
prediction will be present for all weapons. If bc is first copied to the
vector and bc::better_attack() is called only after that, the copy of bc in
the vector will never receive the damage prediction result.


Former-commit-id: b0555b0f27ff4ca89d31a9e28096212ca9e07923
Former-commit-id: c9e0eb87c4ccfaf8ea13b3d30437d0398dd994de
Former-commit-id: 4aad122ea6519505924fea1b53192748b5feee60
Former-commit-id: b11d392b28423a69efdbec0ee39e9e2a5d8fd567 [formerly 359d7a1cb60b0935f76722b6b29fe807cbb58fd4]
Former-commit-id: 61a53e75c06ed399c40ebcd456f27814b7c4b404