Add new event handlers to a staging area
This reverts the semantical change in commit 61ccf2f that event handlers
registered as a result of an event received that event.

Compared with commit 3d8c29bc, this is a lower-level way to fix the ;
character appearing in the command console. This commit has potential to
fix more code that assumes that an event handler can't receive an event
that caused the handler to be registered in the first place. On the other
hand, it is no longer possible for event handlers to receive such events
even if developers want that.


Former-commit-id: 592ea0bce3911ed7ed76933078732bc24f4c0ea5
Former-commit-id: 14919e393f6dc3a197374dcc37a06df3b49ace30
Former-commit-id: b4ece7f26376da281f1f0ec876ab72026b4a57f2
Former-commit-id: b132263ff2551c9eb28bc9d0b8c969d5b3114ebf [formerly cbdc944d22d611b9ac0f03c5f7747ce2dc573f47]
Former-commit-id: 6b6fcef6e776cbda0eda59072c2f4f6bf78c2054