Fixed crash caused by failing to remove an event handler
If code attempted to remove an event handler from an event context whose
all handlers were in the staging area (especially if that context was
freshly created and events::pump() hadn't been called since creating it),
the handler wasn't removed. This resulted in the handler remaining in the
context, but believing that it wasn't there. If the handler was later
destroyed, the event system still retained a pointer to it and attempted to
call handle_event() on it. Depending on the compiler, it could cause a
"pure virtual function call" error message.


Former-commit-id: dbc099f194fa0cc845b7af84c67f4d7acaa56555
Former-commit-id: 7d16985c27872a7b6c8b39f9115ccb251c7b1316
Former-commit-id: 314a5ab72dac796375c528e96404d17559820e50
Former-commit-id: 002686b2cfa1910dc0992a5eccc581b0db322afd [formerly 5df9a54b3c2dea259be9282a2bb2dbeb4bdda768]
Former-commit-id: 1ff3cf86d5b62fa89221cd9327ee5ff327bc1bcc