AToTB S02 Agressive kidnappers
This combines two changes.

First, it uses the sighted event instead of a moveto tripwire to find the kidnappers.

Second, instead of passively waiting for a target to come into range, the kidnappers behave as more normal AI, but ignoring keeps and devalueing villages.

This is done by moving the kidnappers to their own side, which remains hidden until they are discovered.


Former-commit-id: 3f9f0f099ed83e6a91c81e47491871946d538e42
Former-commit-id: 89cef115b750a235bcc6e67cf53ad2749d48decc
Former-commit-id: dac0fe7eb5683988ff73b94a466b9a7a95665744
Former-commit-id: 9dac529bbeae0350a7847d26a97b5e6156a7cac4 [formerly 3419b6317e3d840f0f0a406cd57b759c87e0dd0c]
Former-commit-id: 94d8d666456635b61c0dd75a20214d46656e3f7e