MP Create: improved handling of campaign difficulty selection
Instead of showing the difficulty dialog in post_show and setting the retval to twindow::NONE if the
dialog was canceled, we add a custom exit hook that fires when double clicking the games list or pressing
the Create Game button that shows the dialog instead, and returns if canceled. This is much cleaner.


Former-commit-id: 2cebac2e627d476922632a962bfad400d023e2cf
Former-commit-id: 7d051192388e61686d33239f3fb1f56a78622394
Former-commit-id: 6de4fc0022359e2339aa3c066328c2fc417e2c38
Former-commit-id: 76459f8ca2db1d1e4c167ec3d3fbf739885f7296 [formerly 57d8c877d3b35cf20d473d7134836a8c6b9290a9]
Former-commit-id: 14deafe4380f2ffbadc9424372526615e2591a78