TSG S06a Fix bug: Units in poor shape
When we next see Sir Gerrick and the units he has on the map will be in S09a. While Sir Gerrick will be fine, the units will appear to be in poor shape in the recall list. Let's heal everyone, first, so we're sure everyone is in fighting shape when they return.


Former-commit-id: 0442ee016c6f6445c33701b988db7427c0b67010
Former-commit-id: 9f19050ad8691cada5c7d0a1ce78d763aedf179f
Former-commit-id: 56cac4f29db46c7328bf59c011537ce9e80d1673
Former-commit-id: 8cbf9ac41c6a48c63e96dfd305d6700b033bc6fc [formerly 9a217be9f3ab926f9a282e18653c1b12b5ed1c25]
Former-commit-id: 84d53c2463890688148ecb04dc35cc2a250a3a5a