TSG S05 Let Urza Afala roam free!
There is no reason to stop his moving about. Separating his sighted event and finding the bandit encampment eliminated any problems with his movement. This was a todo comment for when this was fixed. Actually, I think the real problem was the inapproriate sighting of the bandits when we found the lich.


Former-commit-id: 72e3a734ff92cf78b7078ad33f5be1eac5c37cce
Former-commit-id: 8598158fc0351c6f3f1d745a6399b88e65114372
Former-commit-id: f72918215480f9f576ac0197aabcac4d64e9f211
Former-commit-id: 8bf9a06b3069851c44ef2005084fa1d8d4090a3b [formerly 87d87e25f6d529737a6adb22cdc260e3a4666cd2]
Former-commit-id: 75d1ef1046219e69c61b0b526d93779d6452ae04