TSG S01 Fix bug: Urza Mathin may die first
In the unlikely event the player manages to kill Urza Mathin before visiting the river fort, Sir Gerrick will not appear. This is bad for the victory conversation and worse for the rest of the campaign.

I thought long and hard about how to fix this and decided to take my cue from the initial objectives which show two items. So, I made it so you have to do BOTH to win the scenario. If you kill Urza Mathin first, you still need to go find Sir Gerrick. I only needed to adjust one line to make it make sense, and it works better than simply having Sir Gerrick pop up out of nowhere.


Former-commit-id: a0e02337c91bdbbea4735809f37a147cc80f2c28
Former-commit-id: bc8f06078a4478b99b51670e696146e4f2b69139
Former-commit-id: e5ecb0793cc5eed8fe1d847c0fa8699181783f70
Former-commit-id: 50e291b762bc8d79207410fd26328a49e785b7e9 [formerly d317400ad819892a69f9807b228f67193b958388]
Former-commit-id: 801d7527473a3c2ee90978c90a1c2175e621199e