Menu Button: increase min width
This serves the immediate purpose of making the Resolutions Menu Button in Prefs equal the size of
its dropdown, but it also ensures there's more space for lists with small initial entries but longer
choices. Obviously, horizontal_grow = true could be used in any dialogs where this is a case, but this
provides a safer minimum should that not be desired.


Former-commit-id: 9dcd514b7384547c030b49cac2f03b691f1b7c4c
Former-commit-id: 860010e06eb69cfd78a2c8906d712f21e95cb854
Former-commit-id: 641fb946f896f06b347691e449c892ce2054d9ee
Former-commit-id: 968680b068e21aff4207944a175ab40c043cdc02 [formerly 124250ba02d92c4716b31af0420802ef5520749f]
Former-commit-id: b1f18c63eb88dc0ec5b2cfd6f9fef62972a9d665