Load Game: display all human-controlled leaders
This also includes a fix for leader images sometimes not being shown in the dialog.
If the save_index was deleted or altered at a time other than when creating a save,
the image path would be empty in the game load dialog. This ensures the image path
is only saved in binary-path independent form if the path isn't empty. In that case,
the plain image path from the savefile is saved. Either way, the image path is checked
in-dialog, first plain, then as binary-path independent. If both fail, a fallback image
is shown.


Former-commit-id: f61355e2f603eb185923c27a1849dacedfca8a64
Former-commit-id: 084842a6664ba8aec5c3817360155e5f9d860ab5
Former-commit-id: 87dbd5138ab7a31e53fba021b3b4130c54cf6b1d
Former-commit-id: c6af9949bf7c386641e52d643420e069650d5396 [formerly 31b374db5d166d1f8e03cf4a3d30a00ff10ba68c]
Former-commit-id: ee73f41e68493da6956fad3e4f3418510f5a3b29