Load Game: fixed game list being empty with long filenames on small resolutions
For some reason, without horizontal_scrollbar_mode = "never", the game didn't utilize the
window scrollbar and instead simply failed to place the listbox content grid. This resolves that,
as well as reducing the the filename truncate size so names fit fine on 800x600 without said
window scrollbar. Also, it ensures a horizontal scrollbar won't appear in the leaders list.


Former-commit-id: 75c7c63bc926a1608d0846a8a8facf0f36846d05
Former-commit-id: e4c2a89a9eab987a1b76d75cf228d8d69b91f2be
Former-commit-id: 85f82cabd20768b835b15d33234b30becad78d80
Former-commit-id: bb0b5925bb55ffd5d00c9edada1ac9c4478d8ace [formerly b12a2267ab3e8be226e1adb003b4c4842064f9ae]
Former-commit-id: 8e8e9432552bbf5ec3765d6f2838efe25feb9075