Fix two bugs causing multiplayer tests to fail
After finding a game, in some circumstances the join script got the
leader select dialog after only yielding once. Because the loop that
checks when the dialog appears was a repeat-until loop and not a while
loop, it was always run at least once, making the script yield even when
the dialog was already shown. That caused the dialog to close with a
"not shown" status that resulted in exiting the game.

MP tests run two games, not one. A MP wait screen is shown between the
games to the joining player (but not to the host, see
src/game_initialization/playcampaign.cpp:352 ). The join script ended up
quitting between the games. Fixed by simply duplicating code to make the
join script play two games.


Former-commit-id: 3fb4a0f41f27bc8f0682214f25a98aaf68c2da05
Former-commit-id: 7e032c95d94e72e5d6edca551c41eb28e7654937
Former-commit-id: d6b0402fbcd865c7422e76aca9b9db54cef633aa
Former-commit-id: 4d3521e0d5e37480cbc7fc321343e8e884c0c86d [formerly 8357eabd711e1ec1207c1f18e322fa9da2fd87fe]
Former-commit-id: 59d0f67ee8602c7ca6b2728a869e9d709c01ab90