Fix bug: user_team_name is wrong
When pressing End Scenario and proceeding to the next scenario, the user_team_name shown on Menu|Status Table is incorrect.

If you Load (from the title screen) the saved scenario-start, the correct user_team_name appears.

This is caused by the random-shuffle feature for multi-player scenario. This fixes it for Single-Player campaigns and tries not to break it too badly for randomized Multi-Player campaigns.


Former-commit-id: 70c47f639d05256940737287f2d344cb1eaa6848
Former-commit-id: 65e8b3c939fdef8999cf3735847582b430be1d15
Former-commit-id: c815d5271a86d185b50f9c9e91bf1542a0b7af3b
Former-commit-id: b1c8d114528d44b92e4bacae5fc83101fc182df6 [formerly bafdb7f4ebd23fe8855b65e34a255fefeb5e2c5e]
Former-commit-id: 002db5296f95e61b3b981cb18a7f0d67907cf6e9