Default AI: fix bug causing units with max_moves=0 not to attack
The support parameter in the attack evaluation quantifies how much
power the AI can bring to an attack hex. This is supposed to include
the attacking unit itself. However, in some circumstances (such as when
the unit cannot move for one reason or another) this unit is not
accounted for. If no other units can get to this hex either, support is
(erroneously) set to zero, resulting in an attack score of zero, which
in turn results in no attack. Skipping the multiplication of the score
by support in this case does not affect comparison with other attacks
here, as this is by definition the only attack possible on the given
hex.

Note that the '!is_surrounded' conditional is also covered by 'support
!= 0' and can therefore be deleted.

This fixes bug #23720.


Former-commit-id: ccdbf9a847b997d7fce59176fc3c5d5749320532
Former-commit-id: 781036558e8448e79aea0a06882af9dbaf5377c8
Former-commit-id: a4a9e8ca415de5fbdad4763409510bcd8b05be43
Former-commit-id: 0d814b57c88d42801a97e6a9e94f1e0d4030289b [formerly 9ca1f71b5e0ac17a0906e6153c868afbd4c57acc]
Former-commit-id: bce081206e78c4a23f162012504213bc0d959474