Fix crash in GUI2 multiplayer test
Because the join script attempts to join the game immediately after it has
been found, it tried to join the game before it had been added to the
games_ vector (from games_by_id_).

In addition to that, the check for if there is such a game had an
off-by-one error.

The two bugs, in combination, caused the game to index an empty vector.
Even if the STL implementation doesn't immediately crash the game, it's
going to crash later when it tries to interpret garbage data as the game
information.

Fixed by making the game update the games_ vector immediately after
fetching games from the server, and also fixing the off-by-one error.


Former-commit-id: 9835846cd8966235102f99fed27e8a718c7e36cc
Former-commit-id: 5c52cb6631a55e2eee8934d099c752afba382356
Former-commit-id: 315c7830db36df49e6303a9e26138afaa4b55fce
Former-commit-id: c3ac4c3d4a3fd2c0c09866dd6046ade38cfc5320 [formerly 5af492eb4178d0a119d44abe5fda6b10a26da403]
Former-commit-id: 533985a059322a82c72d6a08dfe5de43c39b9a43