HttT S19a Fix bug: Paladin can use sword
An earlier change from blade to arcane for the Paladin's Holy Sword made the Flaming Sword ineffective and he was removed from the list of unit types allowed to take the sword.

This re-instates Paladin by changing the effect filter. I checked all CURRENT units and minimized the filter so the Flaming Sword only effects the swords and the sabre Li'Sar uses.


Former-commit-id: 6485f28ea3a5ba8cf050d3f808c30245fd49206e
Former-commit-id: 79d037743475070ee2454c69d52c03c2512269c5
Former-commit-id: 3e9063b98ae81709334bbcfd3659e75fc6fca538
Former-commit-id: e9d89c4a830e60c2f61370d5df0bc5aa137c6f53 [formerly 477541732c396f3f6c371dc1f2705723b89b70c3]
Former-commit-id: 1477dc1abb69ddc43e1a31ae7e44a6941d67c4fd