fs: Add utility functions for the new file dialog
These include an alternate mode of normalize_path() that enforces the
platform's preferred path delimiter (i.e. backslash on Windows) on the
output, and a function to detect whether a path refers to a root
directory.

Unfortunately, the last bit requires introducing a new link-time
dependency on Windows, against a system library. It's guaranteed to be
always there but it seems kind of a waste. The alternative would be to
hand-parse the string but that seems even more of a waste. And no,
Boost.Filesystem can't do this in a straightforward fashion right now.


Former-commit-id: f20f46a9a7a34e7eca4c2264b6a05649e2007f65
Former-commit-id: ddde4e5f0f87d3d9b00838b10b5594c5008f9e4a
Former-commit-id: bb54818b1ff4adbcff8b855dc390c3f43b29262b
Former-commit-id: 252bb65cea0a433a728e340e951d45dde03add67 [formerly 8d6751a87a723317e1f66a5d6454dd03e8406fae]
Former-commit-id: 5163bcb732100f1bf47ff9e437df27440b6145ab