Honor special locations in terrain mask
A [terrain_mask] is simply a map. It can be smaller than the map, and can be offset. As a map, it can give special locations which can be used to position units, and queried by WML. With this change the names are properly applied to the underlying map, offset relative to the [terrain_mask]. A name given in the [terrain_mask] replaces that name in the underlying map, regardless of the original location. A map file can define only one name per location. A name can have only one location; but a location can have any number of names. To define additional names for a location, stack [terrain_mask] maps giving each. Once placed on the map, a name cannot be removed.


Former-commit-id: bf140255345434ddcf4f220415794051d3dd5011
Former-commit-id: 5c76b5ecc7bf381f03a26d604daef91e201301d5
Former-commit-id: 3c14c0ab84cd1987f028334008562b6b4314649a
Former-commit-id: c5d459653ff377c584e27186ef33101a297cafb5 [formerly 352071a8f50877b3f12d19976ef4f41dba3c9a51]
Former-commit-id: faf3ecc474518d4c52dc2a382e58b480bb6828bc