Change most vector<bool> to dynamic_bitset
A few cases were skipped, most of which involved 2D bitsets (eg shroud)
or uses that could not be easily translated (eg GUI1 multimenu).

In a few cases, code needed to be updated for the different API that the
dynamic_bitset provides, but in most cases, this change is entirely
transparent.

There are probably a few unnecessary dynamic_bitset.hpp includes.

One use of vector<bool> was replaced instead with std::bitset, since
the dynamic size was not required.


Former-commit-id: 3dc09393358e27bc7191850e0c2fe5d6469f3bfa
Former-commit-id: 56c5c59c78e1bf2a560f3d66ba172660b7d91760
Former-commit-id: 6374efdf31196be1a688edfae3203f2b6794f69c
Former-commit-id: 2d751c6dc6b2d2be6a1d7197e90b15a32acd59b8 [formerly b731f61abbddd3814bdd61b8eac10d92ca0f595d]
Former-commit-id: 568d968f82f5c4eb4a5ad5aa2dc03f7f42259710