Check for occupied tunnel exits at teleport map creation
Previously this was done in the move execution, which resulted in:
1. own and allied units on tunnel/teleport exits effectively acting as
ambushers and
2. unit reaches and shortest paths not being consistent with actually
possible moves.
Both issues occurred even without fog being active.

The former was unintuitive at best, while the latter could cause
serious problems, for example with AI code. By moving the check for
units on tunnel exits to the teleport map creation, both these issues
are taken care of simultaneously.

Only visible units are considered. In the presence of fog, enemy units
and units from allies without shared vision continue to interrupt the
move as expected.


Former-commit-id: 71984eb80bc099bb8388016bab3b1a07a93e2c9f
Former-commit-id: 40619e62c78c73812a4b3d34fd82fedcd1127282
Former-commit-id: aaa4421876b7bad04e58789cbccc28347b315fce
Former-commit-id: fab1b5ae89783de21dd41df2d8aa513372fc615f [formerly addc4f725961c6b29f69227bfb3d07e09ed1ce9c]
Former-commit-id: 7e2cddf9840153549174d213e7ee88f31e403c56