Fixed bug where entering Preferences would clear saved password
After refactoring Prefs to use dialog fields, set_remember_password was called every time the dialog
closed. prv::password wasn't initialized at that point, meaning the password was always set to "".
This fixes the bug by explicitly clearing it if false is passed instead of setting it every time.

Previous iterations of the Prefs dialog only called the options setters when they were selected, meaning
this never came up as an issue. Toggling the option off would clear the saved password, and toggling it on
would leave it empty until it was re-entered.


Former-commit-id: 7dc8a3916efd3aa657cec0cb1fc4fac1742034a8
Former-commit-id: 27ae09a553b3545b1107f75edc60a9f4d7b5353a
Former-commit-id: 8cae9a997fe2236fe26cb25430a3343b7ea46059
Former-commit-id: 62075bbf451388a896b9513b7a520e1e6ed45050 [formerly 98eb504d9024a213f65929b83d29d721168c4239]
Former-commit-id: 45fbd9d3084f998d4975abf24ace1108d0f31fc6