finish moving cavegen (cave_map_generator.cpp)  to lua.
This map generator is used in HttT 17 and SoF 4. which needed to be updated aswell.

With this we finally have a lua generator that is used in mainline and that
can be used an an example by wml authors who want to write other map generators.

Unfortunateley there are still umc add ons in 1.12 that use
scenario_generation=cave so i didn't remove the cpp implementation yet.


Former-commit-id: b581bebfebdb1b32a0cedd7e4b67dcd7aecfa61e
Former-commit-id: 529d515c129e47a2e48805db731a67365b27afd0
Former-commit-id: 5b17978d7fcaa2f77cc053bd97f5a6cc2c1f43e5
Former-commit-id: e316359bf77874f063ef7d6765a4222aec231fa6 [formerly 72f58839e7eefa514befe1f2526a254a0a5a1a48]
Former-commit-id: a5da62b7896c9fa2a827db37358ad6e31a2092d0