Remove asserts.hpp
This file seemed to be some sort of reimplementation of <cassert> except
with some "feature" of trying to force the debugger to hit a breakpoint.
However, any decent debugger already hits a breakpoint with a standard
assert(), and this file really only served to make it harder to see the
actual error when using a Windows console.

Some instances of these macros were replaced with assert(), others with
VALIDATE() which instead throws an exception (and possibly puts up a
dialog, if the exception is caught before main()).

A few other related changes that got mixed in:
- Several error conditions now have a better error message (or indeed
  any error message at all)
- Removed an unnecessary use of std::distance on a Boost iterator range.
- Removed a large chunk of code which did nothing but construct a widget
  builder and then crash; the code was unreachable since an earlier loop
  did the same thing (without crashing), and the comment seems to suggest
  that it was a (seemingly no longer necessary) workaround for some buggy
  compilers / linkers.
- noreturn added to the list of compatibilty C++11 features in global.hpp
  (Of supported compilers, only VC12 lacks the new [[attribute]] syntax.)
- Fix detection of GCC in global.hpp

# Conflicts:
#	src/gui/widgets/settings.cpp


Former-commit-id: 9260a4d06116940b33c63e0c6eb3079e196fd3c3
Former-commit-id: 244ffd3cc59ee893c0b6b5f6759891918837df8c
Former-commit-id: 5a4d3153c03ccfa42b2cca40ae3b183092116094
Former-commit-id: 6a35c8b0e3da8da931f3d053fcd821aa3bf3212e [formerly dd5873bd49f13b4c95c26dafc8c9c68bbf7cc0b9]
Former-commit-id: d5e64b245963e84692df22cc8f2a8e7daa0db842