HttT S06 Logical Thieves
If you have already killed Agadla by turn 4, instead of offering a choice, the three who appear simply join your forces.

This means it is possible to never get the Thieves, and not be able to recruit them. But you'd have to win before turn 4; so it's not possible unless you're cheating.


Former-commit-id: 108b20795d56f4e8764878163809481d92b6340b
Former-commit-id: 031672a1b40c0bb57712caf3879db7d0f2277ed3
Former-commit-id: 6b0df7ea43ea42f045aa5fe8c3690e04188ff58e
Former-commit-id: 1df4d08acf19b42a9002375a1c5655fc4d7fab5e [formerly a0be0f0f874d1558736f1f2ce418247a6f5a722b]
Former-commit-id: ebd0de596b683f7183390b73cb22557ecb1717ea