Tunnels: fix issues when exit hex is occupied by invisible unit
This can happen when enemy units are hidden under fog or have the hides
ability. Either way, enemy units always blocks the tunnel.

Invisible allied units may or may not block the tunnel, depending on
the setting of pass_allied_units in the [tunnel] tag. It is therefore
necessary to get and check the full teleport map during a unit’s move
through a teleport, but with see_all=true (while the path finding uses
see_all=false).

This commit fixes the behavior and reinstates the ‘failed teleport’
message.


Former-commit-id: 115691c36fbf887eae682f329dd86f1a22c2dd41
Former-commit-id: 1c51b3bd7a64196118dca80a7353b4470dc76698
Former-commit-id: 3d1c1db9a54a2b73990f3a6aa1712c3f0bd8633b
Former-commit-id: 59f1c3fda4809c51a9b2059e037f173fd1eafc4b [formerly d80550f6f05502f0d4ded2112f90254899df9c6d]
Former-commit-id: d958a5f346ad851d9011429f7418bc6bfa55c4f2