Refactored twesnothd_connection member out of lobby_info
It's only use was to pass refresh_lobby on process_gamelist_diff fail. It makes more sense to
delegate that to the caller - in this case, the lobby. This also means the chatbox widget owns
a twesnothd_connection pointer so the chatbox can be used in situations where no server connection
is present.


Former-commit-id: e475870cb3a6de15268144e9a7c2f7c857d099cc
Former-commit-id: b382cce1034079c3889ddaeaf8e44be66a30ae53
Former-commit-id: 1acb1c9bd19e81e565d0718f77e99df03be92077
Former-commit-id: 32bd4929f9be3b6140b31d4e1f5bdd85a98d17c8 [formerly a18b0ff30baad5e5279b9d73f52807245ffb1912]
Former-commit-id: 6c1891d53648afa225f996d874ba2e0eb12706f6