Uncomment re-selection guard in tgroup::set_member_active
Without this, multiple selections could occur in the Faction Select dialog. However, somehow the
next option immediately to the disabled one gets selected instead of the first available selection
as expected. While this fixes the FLG dialog, it might have repercussions for the Create Unit dialog.

It's possible this issue has something to do with the ordering of the map. Will have to investigate
further after this.


Former-commit-id: 4b8c29532c5432ca951579bc7c4f3527b2d5c3a7
Former-commit-id: 1b22a795c87f45be7cbc57d57b3fd1dcdb2a72e6
Former-commit-id: 2171f6e2709b277893db3b5b9a0c6a7cdeb5b2fb
Former-commit-id: 2687f46da0918ae4ec4888eba5ba9c2997650ab4 [formerly 03adad9bd722a32b4340100e9aa1511e5b86b185]
Former-commit-id: 9ef42325cfcec9a0fe98352bd9ad5a11248bc1d7