Fix rare memory corruption in combat simulation
It occurred when a fight was simulated for a
* slowed unit
* that can level up after the fight
* using Monte Carlo simulation.

Monte_carlo_combat_matrix clears the matrix in constructor because it uses
the matrix in a different way than probability_combat_matrix. Clearing was
unnecessarily thorough and marked all the rows and columns as unused.

If the unit was slowed before the fight, no values were ever placed to
plane 0 (neither unit slowed). Thus, all the rows/columns in plane 0 were
still marked as unused even after the simulation.

After the Monte Carlo simulation, the simulation code considered the
possibility that the unit may have leveled up. It does that by moving all
the values that mean "combatant B dead" into the last row of plane 0, that
means "combatant A full HP and not slowed".

The code that moves the values assumes that at least one row and column is
used. If none are, then it dereferences an invalid iterator and corrupts
memory.

I fixed the bug my making prob_matrix::clear() leave row and column 0 as
used.

After applying this fix, I can't reproduce the assertion failure @mattsc
reported or the hang @GregoryLundberg reported. Those problems were likely
caused by the memory corruption.


Former-commit-id: b86511564e6d75f630b2871025e563b5d1e0dc86
Former-commit-id: 88a94edf9a8539a77e8386865a8832f89af76190
Former-commit-id: 0eb29151bdd0b9308add578a1f9d8484742edc2d
Former-commit-id: cd8346c23161988f555e326f35c9e94841dcf3bb [formerly cbee7e8496976c3ad0dbf50207f178405ef4470f]
Former-commit-id: bf7041dc22c1f06f1468e56d1afc94e9e50d453f