High XP attack CA: fix rare case of AI choosing wrong weapon
Because the attack simulation during CA evaluation is done with enemy
XP set to zero, the AI previously sometimes chose a different weapon
during execution than what was used for evaluation. This resulted, for
example, in archers sometimes attacking an enemy without ranged attacks
with their melee attack when aggression=1.


Former-commit-id: 59e8025c3d76a21fae9c98d0c0c338cdfa6bb683
Former-commit-id: 7bd33e2e9a1719c5ccd5347e091e727e65d203b1
Former-commit-id: c1c07c7d0e71db6e2208fa95f406687f2b851140
Former-commit-id: 0c3178985902dbb28fa155b732b29597c9b15d35 [formerly 60479dbf31710eb78a27822450d7dc662924e859]
Former-commit-id: dfd37958e261770f806331fb6af5d5004e384cbe