add woptipng png compression script
An improved version of the wesnoth-optipng shell script.

Supports multiple jobs at a time. ( -j )
Accepts a threshold for discarding images which were only reduced by a few bytes. ( -t 15 will only save images that were reduced in size by at least 15 %).
Niceness can be set using the -n switch.

Usage:
cd <directory>
woptipng . ../some/dir  ../some/some_file.png

The script should achive better compression ratios than the previous script.
The previous script was running optimizations in an unconditional loop and, at the end, did not save the image if pixel colors changed.
This implies if *one* of the crushers did a bad job and changed the images content in a bad way, *none* of the possibly good optimizations by the other crushed would be saved.
Woptipng checks the image every time after a crusher touched it and only reverts a single optimization if it was bad leaving possibility for the other tools to still perform a good job and reduce size of the image.


Former-commit-id: 6cb97229c7f829044a656d4baa073a368808f6b5
Former-commit-id: 13d3508b43b58880370a7b06b64190bfc43e24fd
Former-commit-id: ef9f854b95b8b0106ecc396f048b7ace68c13da2
Former-commit-id: 744ab58bfab3a85199474d117e003584d95f4460 [formerly 3d03187476b0e24c95f8d599ce2b60ca1207ec61]
Former-commit-id: a0702b4a644eae153a675b8c3a6827762afd1e45