Add a "simulate lobby activity" plugin
This is the first step towards creating artificial load and testing if the
GUI2 lobby still performs well under load.

I also implemented ability to leave the game in the GUI2 MP staging dialog.
Previously the server continued to believe that the player remains in the
game when he/she closed the MP staging dialog.


Former-commit-id: ac1315e1efe158612b5f1cb27a9042e08bdaa13c
Former-commit-id: be7f56f2a7b9024e94741a02a71c98bad421e229
Former-commit-id: 96081c1fe3429b09468190320445e96aa2daf211
Former-commit-id: bc5ef091ff897457d1cd80d95044c45851d810ab [formerly 8736a154a29a68039d75f42b090b1c719f91a8ac]
Former-commit-id: 617aff9dcf2c05d9caa33c8a10c3a0aaa9a75fd8