MP Staging: refactored handling of settings changes
Instead of immediately regenerating the side configs and pushing them to server in every widget callback (which
is especially expensive in the case of the gold and income sliders), we push updates to the server when the
network callback is called if the state_changed_ flag has been set.

This also cleans up some code from the previous attempt to make both the host and clients utilize this dialog.

Additionally, this fixes gold/income changes not being sent to the server.


Former-commit-id: 015eedd2bb294b6a2671c75f94f38ef333071ec6
Former-commit-id: 7641eff9fd5630c83d7dbd67e1cb6a0da90ef8ed
Former-commit-id: ce3707f6339e391eb45caba27700057a0f6a3df7
Former-commit-id: a311ed6e37ed23144b7e93b1c27660eef25c5369 [formerly 1af62dfd54bcc77e890be46978394036d0670355]
Former-commit-id: a8bbaab9fb584961857333289531bc181685b831