Experimental MP lobby: scroll to selected game less often
The lobby calls tlistbox::set_row_shown() every single time there is *any*
change in the gamelist. Most of the time, the visibility of rows hasn't
even changed.

Tlistbox::set_row_shown() calls content_resize_request() to determine if
twindow::invalidate_layout() needs to be called. Content_resize_request()
unconditionally calls show_content_rect() that checks if the selected item
is visible and scrolls to it if it isn't.

With this commit, set_row_shown() returns early if visibility of rows
hasn't changed. This reduces the frequency of scrolling.

This commit isn't a complete fix, since the selected game will still come
into view when, for example, a new game is created. However, at least this
improves the situation.


Former-commit-id: 1250077f7a7b6af5db6fa6813568ea8aeb238b6d
Former-commit-id: 6af43ad7e6d48276a1a917d9b9ea043ab6c23576
Former-commit-id: 2a5ddd680a01f024dd36039b881506d59a2f15f0
Former-commit-id: 19cdec740a0c1de5c3d4de58317cc7735df1c80d [formerly 2ad9f24b6ac515c35b047adfe65cfd859c468a48]
Former-commit-id: 8dc665e3774b00bdc9807b8f2acfa436bf7b1590