Hack: store era addon id in mp game settings
This is so lobby games without the requisite era installed can display a name other than the
era's id, which isn't very nice to look at.

Note that I'm not sure if this is the best or cleanest way to handle this. At some point I'll
need to look more closely at the way addons configs are handled with regards to such games.


Former-commit-id: bd7a7c83930a32300a4fee3c9f976759eea9de82
Former-commit-id: f97d3dc063a0afecc653bd33085269f9381a373a
Former-commit-id: 84bdb03d15b711f95ecd339f036dd34ea6f32b98
Former-commit-id: 76015fc7bb1404502267298bc3df83089674db2c [formerly 89f7fe99f04545ff572ac3fd934d1864df5259d7]
Former-commit-id: 23b85c985bca6b0d73734fc0abdccc0f0da11e35