preferences: Make wesnothd search starting point sane, simplify code
The previous logic reinvents filesystem::directory_name() in a
platform-dependent fashion (it will break if the wesnothd path in
preferences has backslashes on Windows, which is exactly what the GUI2
file dialog delivers), and uses the compile-time install prefix
(non-Windows platforms) or the current working dir (all platforms)
instead of using the game executable's directory, which we can reliably
obtain on Windows or Linux, and automatically fall back to the cwd
otherwise.

Using filesystem::get_exe_dir() as a starting point instead allows for a
much simpler and platform-independent logic (except for the .exe
extension for the binary on Windows, of course).


Former-commit-id: 8fa9d27d26b00fb4bf8875eabce576012f154a16
Former-commit-id: 753846942ab0780224247d6aa1ca9d904997faf1
Former-commit-id: eb8e6d282a81ec33a655628e50be7509eab0d271
Former-commit-id: 2cb563617e849fd9d1ffc46f32216fe6de859e4b [formerly 24ac4900ccecdcf79e0f998428db67cfe6c4b903]
Former-commit-id: f90893454cab611ca82c96e80055c8eda4abb44d