gui2/tfile_dialog: Add support for a filename extension hint
This allows setting a hint in save mode so that the name portion of the
filename will be selected whenever the user selects a file with a
matching extension on the fileview, and so that whenever the textbox is
cleared by code (e.g. when browsing into directories) it is pre-filled
with the extension with the cursor at the start.

With this, the dialog is now feature-complete compared to the GUI1
version.


Former-commit-id: b9217f6678d29fc5450a7f2b69a6d3dc8ef7d43b
Former-commit-id: a25046b1e42b912ae218787f8cfbc0596bf925c2
Former-commit-id: 1ade7cb125504e7566b028428ab6693da18a677a
Former-commit-id: 4db6647598ab8ff82c5ee0fd51a48a0c28adbc58 [formerly 5ba06aece38df3d4e00d1027bfb2025154d6f3b8]
Former-commit-id: 33e77df95272b34577aa5733bd003fb19a67149e