Revert "MP Create: basic implementation of saved games as their own category (debug only)"
This reverts commit 2efd168caa59e865a47195970a1a797fcdc143ee [formerly a0e7eafb474acabe66183d1862a8573c99e372c6] [formerly 8cb4153175e7180f4d1192c561b532d889b3f641 [formerly 0c844ee83cef37b62b41a86060061cb8206123d5]] [formerly 7b178d2659593bc1cb79c0f234de173d70805a63] [formerly 3d09541a1f1d349987cce463dd56ce2f18fe8b8b] [formerly 2205331a0fb17efa2c8b2f1b4f29763d6fb43a77].

This commit somehow causes massive lag when opening the Campaign dialog with a large (several hundred) saves.
Somehow, the save_index was being regenerated every single time a new save was processed, likely due to
the calling of save.summary() in the create_engine initialization loop. It remains to be seen if this is a bug
in the save_index code, which seems possible.

This commit was also likely the cause of bug #25146. While the fix for that might have been simpler (short-circuiting
the campaign type conditional with a savefile corruption check), since this commit has already caused issues, it's
best to revert it and revisit this idea at another time.


Former-commit-id: 135c071a6a25bba6168d4b74fbd49a1ed9ca2174
Former-commit-id: b272c0094e20f305ad9a09a7dfb339b349997e00
Former-commit-id: 6601a84b703998765b73607572c760c97628c1c4
Former-commit-id: 507638efd48c375237c836e194669c1af8e9e07d [formerly 1dbb8534e90cf76b5a58351a5341d7750a9ff9e9]
Former-commit-id: 699ffd650d9fe318511cb6dadb4a684de6ec605c