Added a "Draw Number of Bitmaps" option to the editor
This is similar to "Draw Hex Coordinates" and "Draw Terrain Codes", and displays the number of terrain graphics surfaces draw for each hex. It is useful for spotting mistakes such as overlay images having non-transparent pixels in adjacent hexes where they shouldn't, or for comparing the efficiency of different kinds of terrain graphics rules.


Former-commit-id: 1bdcfa19e94c490c83a350fdd3e50851362012fc
Former-commit-id: 1d5ee6854a78b3b40a5ea460556b6a55143f7a58
Former-commit-id: 6904b0c62279f52e1f722000135c50012409dd24
Former-commit-id: 2f690d6ba902c6facb95da6c86f3840261ed2fc6 [formerly 62c243a861a9c5795cc8e4d32373a304720cceb5]
Former-commit-id: cd8c557b5ff12b55391097a856a2173f1cb75778