Fix a bunch of MSVC2015 compiler warnings about hiding declarations
This commit still doesn't fix all of them. I decided to leave alone
cases where variables with the same name are assigned in multiple
conditions of the same if...else if...else statement, and cases where a
scope logging macro is used multiple times in the same scope. In any
case, this commit greatly reduces the warning count on MSVC2015 and makes
new warnings much easier to spot.


Former-commit-id: ba509a69d2e9e0fa13b46acdd018fe82e7a0e465
Former-commit-id: 63e22d091ac450344e74e70aabe0a098897b9b2a
Former-commit-id: dfd29defcf38715c89a093553c96cb86d4817286
Former-commit-id: c8d57a945ca5eeb3fa01f0c8ca0e52cf052ee86c [formerly 045e6a7809439ee32c706ceea317fb43cb0da4d4]
Former-commit-id: 369a46f9ee0d5e2c5a792eb35525b6db4be2cc23