[filter_vision]: fix bug of filter not matching own/allied hidden units
The vision filter should reflect whatever a player playing the
respective side would see. Previously, any unit with the hides ability
active was excluded from matching irrespective of what side it was on,
even though hidden units of both own and allied sides are always seen
(as long as they are not under shroud or fog when vision is not shared,
but that's covered separately).


Former-commit-id: a19dbc8b6a0a932601ec4baba4abbae844620b05
Former-commit-id: 3f976960c2f1e2bca81d798210d01fbc17cee103
Former-commit-id: 4e5233767ad66c0e1ec1d114d39abc05a348c2e7
Former-commit-id: b398dbc9a169b37425e4e0c00dbc7d4a96b2f98e [formerly 84b72b5c090a80520e3b712bc7057e86f04ed323]
Former-commit-id: f75dc8213e0a30a06402b3a51ef565228656517f