fix an issue with text not properly formatted in lua console
gfgtdf noticed that sometimes, if lua generates text or error messages
which looks like pango formatting, it breaks the markup in the lua
console. the problem was that each unit of text generated by lua
needs to be run individually throw font::escape_text. We identified
the problem but for whatever reason didn't fix it back then, and
put a fixme note instead. This is a proper patch that fixes the issue.


Former-commit-id: 8468e367b58639206fcb1b50971522e7a31628d3
Former-commit-id: ad7b6272d27716713864aac180a05d0778336d21
Former-commit-id: 3a6dd7f1cb40c6d083ee28d157c218055b39dbf6
Former-commit-id: 9030d98b8aa9eaf72b944fa92b4514668b6ea56d [formerly 065e99aa1806ac95688d92a9e742c47c38d57d52]
Former-commit-id: 2728e06cbd3014680d80212901af0a133ff63aa1