add code comments and minor fixups in lua_kernel_base
adds code comments for a number of method declarations in header,
and in the implementation file, prefixes many method calls with
`this->`. This convention disambiguates whether the call is a method
of the class or a function at global scope. On some projects like
gcc, it is part of the coding standard to do this to avoid subtle
bugs -- I have come to believe that it makes your code more easily
readable, even if it is a little more typing.


Former-commit-id: e30ea6cd1363c954c875407bf2bcca3814d14384
Former-commit-id: 0c6b04d67271bd29c55a2bb154db55dbc9ff2adf
Former-commit-id: 0fc9815110c1b10dfcb88907288fab2f8dec6563
Former-commit-id: 63e935ba35b3edd1b7cf87c2576d5f5ee52ecbab [formerly 9905839de4688e9213d5851b4822300b8f3c085a]
Former-commit-id: d4e6785c9dfa6892dff70acffedc5301ba749c0c