campaignd: Only quit to SIGINT and SIGTERM after a processing cycle
Otherwise we run the risk (as seen with the 2016-10-07~09 downtime
incident) of calling exit() while we're in the middle of writing WML
content to disk, in particular the add-ons list and server configuration
file. Trust me, no-one will be very impressed if you make them sort out
the mess that this kind of thing leaves behind.

Ideally we should atomically rename a temporary into the destination
file for each WML write we do here to avoid similar ugliness with
signals we don't/can't/mustn't handle (especially SIGKILL). It's always
best to stay on the safe side and assume that a botched write could kill
people.


Former-commit-id: e0d93b88b5a0bd52b021a8ed567faeb63f47112c
Former-commit-id: 826c0d2a6c09193102300323bec1bbdc63367f69
Former-commit-id: a7543cd0a1b2a3f77330817894055fa47c674c27
Former-commit-id: 48fe07d07e7bc7f309518d01ec1934fc68db4492 [formerly ab2ab63c5fb41f0a3f3929680e6811d0e2c30065]
Former-commit-id: 1569a304449ca572172e0d2d28dfec78abc3be23