ai_helper: correctly and consistently deal with invisible units
Previously, some functions saw invisible units, while others did not.
Moreover, this behavior could be triggered on and off for some
functions, but not for others, and the parameter to do so was not the
same in all cases.

Now, the default setting for all ai_helper functions is to ignore
invisible units and this can be turned on and off with the same
parameter, viewing_side, with the same syntax and meaning as for the
built-in functions wesnoth.find_path() and wesnoth.find_reach(). This
means hat it can be passed directly between those functions and all the
ai_helper functions.


Former-commit-id: 9500695e1e328f7747d65474315d4d18c3f353c4
Former-commit-id: cb980432b0d1de68f991f978f47817f220588fe4
Former-commit-id: f82e8315517a56e8c4438cd21d5fc5d53ef3e1e0
Former-commit-id: 348f27608211d44d00252004357e7572bb79aa0d [formerly 27901e23816918c29a9fe6681f9cbdc2122bb28a]
Former-commit-id: 579644239ca0cc00a886b7facbb9688d23468a32