campaignd: Add wrapper for atomic commits of crucial files
As the 2016-10-07~09 downtime incident shows, it is paramount to take
further steps in guaranteeing that the server can't corrupt its own data
files (especially the add-ons database) when receiving
inappropriately-timed signals.

This commit adds and deploys an ostream wrapper that requires callers to
explicitly commit the results to disk when finished writing to the
stream, so that only then the real destination file is overwritten with
the working copy (a temporary in the same dir). This way, there should
never be a situation in which the destination is left in an inconsistent
state due to a signal or exception.

The temporary receives a predictable name right now in the interest of
simplicity, since we are more or less in control of the target directory
anyway. We definitely don't want it to be an unlinked file since that
would make it impossible for admins to inspect and compare the temporary
against the original afterwards.

The code makes some assumptions about the nature of the return value of
filesystem::ostream_file() which will never be broken in this stable
branch, which is why one helper function is in campaignd land rather
than in the global filesystem API for now. This should probably be
rectified when forward-porting to master. Maybe.

Nothing of this will work reliably on Windows but we don't care. There's
only one machine in the world where we support running campaignd at this
time and it runs Linux.


Former-commit-id: df9cad4e53b6015c96c9e7ca53e54a46da7ca827
Former-commit-id: e2ec78b7ad5d130b37db9749de944c3bcc4718e4
Former-commit-id: 9ff54eb7d54095b7da0036d5f625e5d303ac19f8
Former-commit-id: 4873138be7d257416a1715c56ff1301b18467e58 [formerly e95226c24e5fa84d97896869be741fa7f93c4f4c]
Former-commit-id: 7b083d51624b016ff113292a1f809507d20f0e15