Correct the calls instead
Do not change the released Lua source code. Lua is written as standard C, but we compile it as C++ so we can use exceptions. Change the call sites to avoid the warnings about using a C cast instead of static_cast<int>().

NB: The changes to Wesnoth C++ code will be required to upgrade to Lua 5.3, anyway.

This reverts:

commit f6d013404525ca5d8d14061f6b0f27251d7ba1b3 [formerly bba8912db58aceade3e2d0dda674ed69c02774f1] [formerly 470ec3328ae5050ed0b3a52254c4c43e8e8f83ab [formerly 9515261d1ba2ff1acb6831ef7ad5ae6e56327e8a]] [formerly 7a36d5c1b3265924cb024703abcbf5eb39fcea55] [formerly ec7efe67efe4bca8c4d9146d7f95e274c54b7fc5] [formerly 9e1d253bdb256c80c6716d2e3d1320db885f1f58]
Author: Mark de Wever <koraq@xs4all.nl>
Date:   Sun Feb 5 19:55:32 2012 +0000

    Fix compiler warnings.


Former-commit-id: 3b12fc0ae70ca6d8e9bacd804ae15417a261ad9b
Former-commit-id: b0a0148b8deb0c1b8c84704bdf52d6531a8af42b
Former-commit-id: 9615476e1eabf4124f37e8932e2945e010c100ed
Former-commit-id: 856035692ebe9761086c35b0db7fc3a5bed29a34 [formerly 8da451077170d443d6206a7e03ac73a0171bbc32]
Former-commit-id: 61c5f3fd3c38bcf122876be16df86d088e028fe3