desktop: Add functionality to enumerate paths of interest
This allows querying things such as the user's home dir, drive letters
(Windows-only), and game paths such as the data dir, preferences, user
data, and binaries. The results are presented in a format that's
suitable for UI use, with translatable labels used wherever applicable.

At some point there'll be support for listing user-defined bookmarks as
well.


Former-commit-id: 362e70bd8c5649484d873866d4ebf93a25a49865
Former-commit-id: 9f048ebb1e216e8b201228632256ddf3bcbc3751
Former-commit-id: a884282892ae082bcecf5f6141e95390be0d7803
Former-commit-id: dc09e09618d5b0ad3a611fa6e4a01d79060708a1 [formerly cda721d5f04a6082a39f8784201de89674080505]
Former-commit-id: 007965c009376831d6be186f09c66955da8f9b92