gui2/tfile_dialog: Add a bookmarks bar
The bookmarks bar holds predefined bookmarks (as in the stuff
src/desktop/paths.hpp exposes) and allows users to easily browse to them
in a single click. It will eventually be possible for the user to add or
remove custom bookmarks as well.

I might add a method to disable specific irrelevant bookmarks later, not
sure (e.g. nobody cares about the preferences dir when trying to find
wesnothd).


Former-commit-id: 3a7f527c9ecdd70a68117b9207ef6eb2e998e0d0
Former-commit-id: 92238c4018a82808b35cf80f92f821f3a7a8db51
Former-commit-id: 418fa2098e50fdff8a76329d85e657a5f7cb1c2d
Former-commit-id: 27ef8e074b1e95b999f31909b3edc4bc50629762 [formerly c2918a3236df1e0d5374464d70a08ab928a52fee]
Former-commit-id: 0024e2196c42642e083c34fd8858b739fd801f1c