gui2/tfile_dialog: Rearrange some elements
The path label above the fileview now stands on its own without the
"Location:" label since it's self-explanatory for anyone who's managed
to get this far and install Wesnoth. The Delete and New Folder buttons
have had their order reversed so that destructive actions are on the
side opposite to the dialog's control buttons (OK/Cancel), the "Places"
heading no longer has a colon.


Former-commit-id: 9e706a45ee74734190b03d2a74feb1f2b5b68629
Former-commit-id: 90df24de64d555ec93a2bfba25b4cc797105ded4
Former-commit-id: f88b3c9039109d1c2872a8c6626732da815d70f3
Former-commit-id: 81821b2f7c1534b43aadcf9709e30a4cd81e96e3 [formerly 18197bbe394b08b4238455dcdd46108bfeb8981b]
Former-commit-id: 6f3faef401d6a10ebc04a4bb03bb2028bb010be8