gui2/tfile_dialog: Add support for user-defined bookmarks
Users can add and remove bookmarks from the end of the bookmarks list.
These correspond to the currently viewed directory and have its name as
their label. It's not possible to remove hardcoded bookmarks. A bookmark
may only be removed after selecting it.

In order to ensure that users will be able to remove their own bookmarks
if they match predefined paths, the bookmarks bar now matches items from
bottom to top instead of the other way around.


Former-commit-id: af6d9d2f20a055f60fab1322b9d8492a59b2d31b
Former-commit-id: 6772de4dc75cb1893280184707562decc2326658
Former-commit-id: 0914598ed8e021c57c4d2978776e9d9e91daf860
Former-commit-id: 567c2c6fc8e680b6c5691294c330e61311ea2cb4 [formerly e0f1841c18a8df1314d8316202bc80c1f1a9e5b9]
Former-commit-id: cd6052bd37c7fb55ad59d8a11c3c9519de7d6546