Game Config: added default_color_list= key
This key contains the color range ids of the standard colors in the default order.

Previously, classes such as the connect_engine would have to cast ints to strings and browse for the numerically id-ed
color ranges. Consequently, since there are only 9 ranges identified by that method, 9 became a hardcoded number of
colors in certain places. This key allows for greater control over order and content of the default colors.


Former-commit-id: 80930446a4bdb68232acef7fe91bf993eedffdeb
Former-commit-id: 421735143b70c01ec108be6615ba12c0b155d08a
Former-commit-id: 7f723173003155d67b43483e81476aeb593dfd7d
Former-commit-id: 73c5ddc8c92db2b6c0a5e1f12c0eef4fadb89cf9 [formerly f6b21859a3dc940948ad5bfb012bdeea979fd133]
Former-commit-id: f62adbd0826c02a308864a352d6ae0274ec09c5d