Connect Engine: refactored handling of colors
Previously, if any [side] color= were specified at all, a new entry would be added at the beginning of the color
options list, regardless of whether it previously existed as an option or not, and the side_engine::color()
variable was set to 0. This was fine for displaying a color as a value in a comobox, but it looked awful
if one actually looked at the list. Additionally, color() could not be used to apply team coloring (either
via ~RC or ~TC) to anything, since it would always be pointing at 0!

This refactors handling to use more dynamic and sensible methods. A copy of the game_config's default color
list initially populates the color options.

* If any value of color= is already found within that selection, color() will be set to that index value.
* If a numerical value is provided instead, that will be used directly.
* If any non-default color is specified, it will be appended to the end of the color options.

This also removes the getters for pango markup for color options. This is outside the scope of this class and
should be handled by any callers.


Former-commit-id: af3e2c6b6bc99bfdb6d6d81c02dc97f1f79e9bf7
Former-commit-id: 08dac42dceaa6c05a6d5593071e3836c55d34e58
Former-commit-id: a739a7dc772c94eaa2dae59d8cfe316d564610d8
Former-commit-id: c75b41c14c7603a817d86f2341230102f6a8d5fd [formerly 233c1aca354ac5d690b8b936845e14550e4d122c]
Former-commit-id: db20d174e7cb4344bd006b9f0e8cba478b0d95ca